Increase nonce dto 64 bit and improved authorization on Swagger page. Also added new test project.

+++ src/InvestorApi.Contracts/IAccountService.cs
@@ -57,7 +57,7 @@ namespace InvestorApi.Contracts
         /// <param name="symbol">The share symbol.</param>
         /// <param name="quantity">The quantity to buy.</param>
         /// <param name="nonce">The nonce value required to detect suplicate orders.</param>
-        void BuySharesAtMarketPrice(Guid userId, Guid accountId, string symbol, int quantity, int nonce);
+        void BuySharesAtMarketPrice(Guid userId, Guid accountId, string symbol, int quantity, long nonce);
 
         /// <summary>
         /// Sells the supplied quantity of shares at the current market price.
@@ -67,6 +67,6 @@ namespace InvestorApi.Contracts
         /// <param name="symbol">The share symbol.</param>
         /// <param name="quantity">The quantity to sell.</param>
         /// <param name="nonce">The nonce value required to detect suplicate orders.</param>
-        void SellSharesAtMarketPrice(Guid userId, Guid accountId, string symbol, int quantity, int nonce);
+        void SellSharesAtMarketPrice(Guid userId, Guid accountId, string symbol, int quantity, long nonce);
     }
 }
